requirements.txt
django-rest-framework
django-filter
django-rest-swagger
djangorestframework-jwt
django-rest-auth
celery
django-celery==3.2.2
django_celery_results
django-celery-beat
flower==0.9.2
eventlet
supervisor
uwsgi